refac : interface 추가
